requirements.txt
pytz
pyyaml==6.0
twm-faust[rocksdict,prometheus]>=1.14.3,<1.15.0
python-benedict==0.33.2
marshmallow==3.21.3